+++ backend/src/main/java/de/floralbite/backend/offers/models/ColdOfferCategory.java
@@ -1,4 +1,4 @@
-package de.floralbite.backend.offers;
+package de.floralbite.backend.offers.models;
 
 public enum ColdOfferCategory {
     FINGERFOOD,
