+++ backend/src/main/java/de/floralbite/backend/offers/models/ColdOfferDTO.java
@@ -1,4 +1,4 @@
-package de.floralbite.backend.offers;
+package de.floralbite.backend.offers.models;
 
 public record ColdOfferDTO(String name, ColdOfferCategory category, String description) {
 }
