+++ backend/src/main/java/de/floralbite/backend/offers/models/MenuOfferDTO.java
@@ -1,4 +1,4 @@
-package de.floralbite.backend.offers;
+package de.floralbite.backend.offers.models;
 
 import java.util.List;
