+++ backend/src/main/java/de/floralbite/backend/offers/models/SweetOfferCategory.java
@@ -1,4 +1,4 @@
-package de.floralbite.backend.offers;
+package de.floralbite.backend.offers.models;
 
 public enum SweetOfferCategory {
     DESSERT,
