+++ backend/src/main/java/de/floralbite/backend/offers/models/SweetOfferDTO.java
@@ -1,4 +1,4 @@
-package de.floralbite.backend.offers;
+package de.floralbite.backend.offers.models;
 
 public record SweetOfferDTO(String name, SweetOfferCategory category, String description) {
 }
