+++ backend/src/main/java/de/floralbite/backend/offers/repos/MenuOfferRepo.java
@@ -1,5 +1,6 @@
-package de.floralbite.backend.offers;
+package de.floralbite.backend.offers.repos;
 
+import de.floralbite.backend.offers.models.MenuOffer;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
